Have all OS's always use portability queue routines.
